test: :test_tube: wip permission TU is failing

+++ apps/server/src/controllers/permission.js
@@ -3,7 +3,10 @@ import {
   setPermission,
   deletePermission,
 } from '../models/queries/permission-queries.js'
-import { getRoleByUserIdAndProjectId } from '../models/queries/users-projects-queries.js'
+import {
+  getRoleByUserIdAndProjectId,
+  getSingleOwnerByProjectId,
+} from '../models/queries/users-projects-queries.js'
 import { getLogInfos } from '../utils/logger.js'
 import { send200, send201, send500 } from '../utils/response.js'
 
@@ -76,6 +79,9 @@ export const updatePermissionController = async (req, res) => {
     const role = await getRoleByUserIdAndProjectId(userId, projectId)
     if (!role) throw new Error('Requestor is not member of project')
 
+    const ownerId = await getSingleOwnerByProjectId(projectId)
+    if (data.userId === ownerId) throw new Error('La permission du owner du projet ne peut être modifiée')
+
     const permission = await setPermission({ userId: data.userId, environmentId, level: data.level })
     req.log.info({
       ...getLogInfos(),
@@ -104,6 +110,9 @@ export const deletePermissionController = async (req, res) => {
   try {
     const role = await getRoleByUserIdAndProjectId(requestorId, projectId)
     if (!role) throw new Error('Requestor is not member of project')
+
+    const ownerId = await getSingleOwnerByProjectId(projectId)
+    if (userId === ownerId) throw new Error('La permission du owner du projet ne peut être supprimée')
 
     const permission = await deletePermission(userId, environmentId)
